change type of code argument in CompressionAlgorithm.fromCode to int from byte

+++ foreign/java/java-sdk/src/test/java/org/apache/iggy/topic/CompressionAlgorithmTest.java
@@ -32,8 +32,7 @@ class CompressionAlgorithmTest {
     @ParameterizedTest
     @CsvSource({"1, None", "2, Gzip"})
     void fromCodeReturnsCorrectCompressionAlgorithmForValidCode(
-            byte code, CompressionAlgorithm expectedCompressionAlgorithm) {
-        // TODO - check if byte was intended here
+            int code, CompressionAlgorithm expectedCompressionAlgorithm) {
         assertThat(CompressionAlgorithm.fromCode(code)).isEqualTo(expectedCompressionAlgorithm);
     }
